+++ fuse/context.py
@@ -20,7 +20,6 @@ microphysics_plugins = [
     fuse.micro_physics.cuts_and_selections.XENONnT_TPC,
     fuse.micro_physics.cuts_and_selections.XENONnT_BelowCathode,
     fuse.micro_physics.cuts_and_selections.DefaultSimulation,
-    fuse.micro_physics.cuts_and_selections.ApplyCuts,
     fuse.micro_physics.ElectricField,
     fuse.micro_physics.NestYields,
     fuse.micro_physics.MicroPhysicsSummary,
